Max Kellermann <max@duempel.org>: use size_t for buffer sizes

+++ include/buffer.h
@@ -1,20 +1,22 @@
 #ifndef _BUFFER_H_
 #define _BUFFER_H_
 
+#include <stddef.h>
+
 struct buffer {
 	unsigned char *data;
-	unsigned int size;
-	unsigned int cur_size;
+	size_t size;
+	size_t cur_size;
 };
 
-struct buffer *buffer_create(unsigned int size);
+struct buffer *buffer_create(size_t size);
 void buffer_destroy(struct buffer *b);
 
-int buffer_add(struct buffer *b, void *data, unsigned int size);
+int buffer_add(struct buffer *b, void *data, size_t size);
 void buffer_flush(struct buffer *b, 
 		  void (*cb)(void *buffer_data, 
 		  void *data),
 		  void *data);
-unsigned int buffer_size(const struct buffer *b);
+size_t buffer_size(const struct buffer *b);
 
 #endif
